Rough design ideas for 3D printed base and parts.
In the past lab Ben and I went over some possible designs for the custom housing for the robot and felt the coffin type design would work best. We then worked out how each part would be attached to each sensor or the Pi itself. From here we devised a list of parts we need to create and what parts we will be creating each week. The first week will be used to get used to the design software by only doing the simple sensor holding designs.