Rough Component Layout + Motor housing parts
The RoughComponentLayout.jpg shows a basic idea of where all the different components will sit on the custom base to the robot.This plan was done by both Ben and I. The motor.stl files are the completed 3D printable files that will create a motor housing. The MOTOR.png shows an image of how the motor parts roughly look and will fit together.