Example of Base+Motors+Casing together
Parts will be printed separately but i put them into one final so you could have a look at how they might fit together. The design is able to be changed. Ben is working on the IR holder but I asked him to change  a couple things so it could fit on to the board in a more natural way.