Some alternate movement code
Although this code does not use differentials yet, the proportional code is a more reliable version of out original. Instead of the sum value it uses the centre position of the line (and we can use the value '160' rather than an estimated max value). This means the camera at different heights shouldnt affect the code.