Final robot with pieces screwed together and RPi attached
All of the parts including the base and casing were 3D printed. All of my parts and Ben's IR sensor fit well onto the board and held securely. The only problem with this was the structural integrity of the casing and the weighting of the robot. I did not make the casing pieces have enough support resulting in the front piece breaking off so now the RPi is connected only to the back piece. The other problem was the weighting. The wheels were very central rather than at the front to save space but this meant it was easier to tip over so the battery will now reside at the back along with the RPi to balance the robot whilst moving.